Move all the option parsing stuff to its own files
